requirements.txt
flask
tensorflow
numpy
geopy
osmnx
networkx
folium
google-cloud-firestore
matplotlib
scikit-learn
google-auth
ortools
scipy